+++ classes/privacy/provider.php
@@ -1,0 +1,48 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * Shoelace theme.
+ *
+ * @package    theme
+ * @subpackage shoelace
+ * @category   privacy
+ * @copyright  &copy; 2018-onwards G J Barnard in respect to modifications of the Bootstrap theme.
+ * @author     G J Barnard - {@link http://moodle.org/user/profile.php?id=442195}
+ * @author     Based on code originally written by Bas Brands, David Scotson and many other contributors.
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace theme_shoelace\privacy;
+
+defined('MOODLE_INTERNAL') || die();
+
+/**
+ * The Shoelace theme does not store any user data.
+ * But if you use CDN fonts then investigate the respective privacy policies.
+ */
+class provider implements \core_privacy\local\metadata\null_provider {
+
+    /**
+     * Get the language string identifier with the component's language
+     * file to explain why this plugin stores no data.
+     *
+     * @return  string
+     */
+    public static function get_reason() : string {
+        return 'privacy:nop';
+    }
+}